requirements.txt
torch
numpy
transformers
safetensors
pillow
image
